Optimize Your Resume for Applicant Tracking Systems
Job seekers typically understand that there's some sort of digital barrier standing between them and the jobs they see advertised online. And they know they need to use keywords to overcome it. But many aren't familiar with applicant tracking systems (ATS)—the software that acts as a gatekeeper for companies and recruiting agencies. In this course, career strategist Jenny Foss explains how recruiters use this software, as well as how to create a winning resume—one that’s not just a compelling read for recruiters, but also optimized for the ATS. Discover how applicant tracking systems work and how to select keywords to maximize your chances for success. Explore how your choice of formatting, fonts, and graphics can trip up the ATS. Plus, learn about the best type of resume for the ATS, as well as how to rework your own resume to ensure it gets into the hands of the right person.