Cognitive Technologies: The Real Opportunities for Business
Cognitive technologies such as artificial intelligence and robotics are changing how businesses operate and the nature of work as we know it. This course, from Deloitte University Press, is designed to explain the benefits and value of cognitive technologies to business leaders, decision makers, and others who want to understand their impact on business. David Schatsky focuses on the "what" and "why," leaving you to craft a "how" that meets your organization's needs. He covers machine learning, artificial intelligence fields such as natural language processing and computer vision, and robotics. In later chapters, David examines the business case for the technologies, looking at practical applications for products and processes. Finally he reviews the impact on workers and the design of work, and takes a look forward into the future of cognitive tech. Short quizzes and assessments help you practice your knowledge. By the end of the course, you should be able to engage in productive discussions with colleagues, customers, and suppliers and help shape the cognitive technology strategy at your organization.