AI in Business Essential Training
Artificial intelligence (AI) is changing the way we interact with our mobile devices, our healthcare providers, and even our cars. In this project-based course, Dr. Michael McDonald focuses specifically on the applications of AI in business, as well as emerging opportunities in the space. After discussing the basics of AI and how it works, Dr. McDonald focuses on how this exciting technology can be leveraged in a variety of business settings. Discover how AI can be applied to accurately price products, make astute marketing decisions, navigate financial markets, and more. Plus, learn where AI doesn't work in the business world and get best practices for applying AI in your work.