Picking the Right Chart for Your Data
When selecting charts to showcase data, many people simply pick from the few choices available in Excel or other such software. This can be highly limiting, and can result in the selection of charts that fail to effectively convey whatever you're trying to communicate. This course can help you think more strategically about your data, and provide you with the tools you need to pick the best visual display for the type of data you're working with—and your ultimate communication goals. Main topics include getting to the key idea you're trying to communicate; finding the right standard chart for your data type; and brainstorming and experimenting to come up with alternatives to the standards.